requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
idna==3.4
inflection==0.5.1
jsonschema==4.17.1
PyJWT==2.6.0
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13